Pre-render RsT math without Warehouse head scripts
Since we are pre-rendering everything through
KaTeX, we no longer need any additional javascript
being shipped to the client for MathJax.

Warehouse would only need the CSS in the head
of the HTML and would offload the rendering
to readme_renderer, thus improving performance
for large documents and older machines.